re-organize thunks
